All problems solved
