+++ apache-config.conf
@@ -1,0 +1,12 @@
+Alias /web_app /var/www/html/src/web_app
+Alias /api /var/www/html/src/api
+
+<Directory /var/www/html/src/web_app>
+    AllowOverride All
+    Require all granted
+</Directory>
+
+<Directory /var/www/html/src/api>
+    AllowOverride All
+    Require all granted
+</Directory>
